third_party/WebKit/LayoutTests/bluetooth/server/disconnect/detach-gc.html

<!DOCTYPE html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../../resources/testdriver.js"></script>
<script src="../../../resources/testdriver-vendor.js"></script>
<script src="../../../resources/bluetooth/bluetooth-helpers.js"></script>
<body>
<script>
'use strict';
const test_desc = 'Detach frame then garbage collect. We shouldn\'t crash.';

bluetooth_test(() => setUpHealthThermometerDeviceForDiscovery()
  .then(() => createIframe())
  // Connect device, detach the iframe, and run garbage collection.
  .then(iframe => new Promise(resolve => {
    let options = {filters: [{services: ['health_thermometer']}]};
    messageIframe(iframe, 'RequestAndConnect', options);

    window.onmessage = messageEvent => {
      assert_equals(messageEvent.data, 'Connected');
      iframe.remove();
      runGarbageCollection().then(resolve);
    }
})), test_desc)
</script>
</body>
